% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_data.R
\name{baseline_inputs}
\alias{baseline_inputs}
\title{Load baseline inputs}
\usage{
baseline_inputs()
}
\value{
A list.
}
\description{
This function returns a list with input parameters used in the baseline
}
