% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_simulation.R
\name{full_ceq}
\alias{full_ceq}
\title{Title}
\usage{
full_ceq(inps, presim, ...)
}
\arguments{
\item{inps}{input file}

\item{presim}{pressimulation file}

\item{...}{other parametsr}
}
\value{
completed microsimulation
}
\description{
Title
}
