% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_variables.R
\name{get_var_nm}
\alias{get_var_nm}
\alias{get_inc_nm}
\alias{get_wt_nm}
\title{Key variables names and labels}
\usage{
get_var_nm(vars = NULL, suffix = "_pc")

get_inc_nm(suffix = "_pc")

get_wt_nm(suffix = NULL)
}
\value{
data frame with variables names

a data frame with variables names

a data frame with variables names
}
\description{
Key variables names and labels
}
\section{Functions}{
\itemize{
\item \code{get_inc_nm()}: Income Concepts variables IDs and labels returned in a data frame

\item \code{get_wt_nm()}: Weight variable IDs and labels in a data frame

}}
