% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_data.R
\name{load_input_structure}
\alias{load_input_structure}
\title{Load Input Structure}
\usage{
load_input_structure()
}
\value{
A list containing the input structure.
}
\description{
This function loads the input structure from the `geo_inputs.xlsx` file included in the package.
}
\examples{
inputs_str <- load_input_structure()
}
