% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_data.R
\name{load_input_tables_structure}
\alias{load_input_tables_structure}
\title{Load Input Tables Structure}
\usage{
load_input_tables_structure()
}
\value{
A list containing the input tables structure.
}
\description{
This function loads the input tables structure from the `geo_inputs.xlsx` file included in the package.
}
\examples{
inputs_table_str <- load_input_tables_structure()
}
