% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_data.R
\name{load_input_tabs_structure}
\alias{load_input_tabs_structure}
\title{Load Input Tabs Structure}
\usage{
load_input_tabs_structure()
}
\value{
A list containing the input tabs structure.
}
\description{
This function loads the input tabs structure from the `geo_inputs.xlsx` file included in the package.
}
