% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_georgia_ceq_app.R
\name{run_georgia_ceq_app}
\alias{run_georgia_ceq_app}
\title{Run Georgia CEQ Microsimulation Tool}
\usage{
run_georgia_ceq_app(
  path_to_presim = NULL,
  path_to_baseline = NULL,
  name_app = "Georgia CEQ Microsimulation Tool"
)
}
\arguments{
\item{path_to_presim}{Character string. The file path to the presimulation `.rds` file.}

\item{path_to_baseline}{Character string. The file path to the baseline `.rds` file.}
}
\description{
This function launches the Georgia CEQ Microsimulation Tool Shiny application.
}
\examples{
\dontrun{
run_georgia_ceq_app()
}
}
