% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_data.R
\name{set_up_data}
\alias{set_up_data}
\title{Set up data file path}
\usage{
set_up_data()
}
\value{
A character string with the path to the `geo_inputs.xlsx` file.
}
\description{
This function returns the path to the `geo_inputs.xlsx` file included in the package.
}
