% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_presim.R
\name{set_up_presim}
\alias{set_up_presim}
\title{Set Up Presimulation Data as a Reactive Expression}
\usage{
set_up_presim(path_to_presim, path_to_baseline)
}
\arguments{
\item{path_to_presim}{Character string. The file path to the presimulation `.rds` file.}

\item{path_to_baseline}{Character string. The file path to the baseline `.rds` file.}
}
\value{
A reactive expression that, when evaluated, returns a list containing the presimulation data with the baseline data included.
}
\description{
Reads presimulation and baseline data from specified file paths and returns a reactive expression containing the combined data.
}
